pyproject.toml
[tool.poetry]
name = "parq-blockmodel"
version = "0.1.5"
description = "A Python package for efficient storage, manipulation, and analysis of mining block models using Parquet files."
authors = ["Greg <11791585+elphick@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "parq_blockmodel" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
numpy = ">=1.25.2,<2.0"  #2.1 max for numba->ydata-profiling compatibility
pyarrow = ">=16.0"
parq-tools = ">=0.3.1"
pyvista = ">=0.44.2"
geopandas = "^1.1.1"
tqdm = { version = ">=4.67.1", optional = true }
ydata-profiling = { version = ">=4.16.1", optional = true }
setuptools = { version = "*", optional = true }
plotly = { version = ">=5.24.1", optional = true }
#parq-tools = { git = "https://github.com/elphick/parq-tools.git", branch = "12-column-descriptions-in-profile-report" }
#ezdxf = { version = "^1.4.2", optional = true }
#scikit-image = "^0.25.2"
rioxarray = { version = ">=0.18.0", optional = true }
imageio = { version = "^2.37.0", optional = true }
imageio-ffmpeg = { version = "^0.6.0", optional = true }
kaleido = { version = "^1.1.0", optional = true }
#htmlmin = "^0.1.12"


[tool.poetry.extras]
profiling = ["ydata-profiling", "setuptools"]
progress = ["tqdm"]
temporal = ["rioxarray", "imageio", "imageio-ffmpeg"]
viz = ["plotly", "kaleido"]  #, "exzdxf"]
all = ["ydata-profiling", "setuptools", "tqdm", "rioxarray", "plotly", "imageio", "imageio-ffmpeg"]  #, "ezdxf"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
sphinx = "^7.3.7"
sphinx-gallery = "^0.14.0"
sphinx-rtd-theme = "^2.0.0"
myst-parser = "^3.0.1"
sphinx-autodoc-typehints = "^2.1.1"
matplotlib = "^3.7.2"
coverage = "^7.8.0"

# packages below here are for plotly and plotly thumbnails
#kaleido = [
#    { version = "0.1.0post1", markers = "sys_platform == 'win32'" },
#    { version = "0.2.1", markers = "sys_platform != 'win32'" }
#]
trame = "^3.10.2"
trame-vtk = "^2.9.0"
trame-vuetify = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
